Improved configuration for server disks.
Note that this is a breaking change - 'image_disk' and 'additional_disk' are now combined into 'disk'.
